does this work?

+++ trolley-arch-chroot.sh
@@ -7,7 +7,11 @@ else
 fi
 # install grub
 pacman --noconfirm -Sy grub
-grub-install --target=i386-pc --force $DEV
+if [ -e /sys/firmware/efi/efivars ]
+then
+    grub-install --target=x86_64-efi --force $DEV
+else
+    grub-install --target=i386-pc --force $DEV
 grub-mkconfig -o /boot/grub/grub.cfg
 # set up root password
 read -ps "Please enter your root password: " ROOTPASS
